Cleanup the send to slack message.
