+++ website/src/components/page/changelog/web.json
@@ -1,9 +1,9 @@
 [
   {
-    "productTitle": "Portals Web",
+    "productTitle": "Portals Web Plugin",
     "pageUrl": "https://ionic.io/docs/portals/for-web/changelog",
     "repo": "ionic-team/ionic-portals",
-    "mdBody": "### What's Changed\r\n* **fix**: Peer dep was not upgraded to cap 5 ([#228](https://github.com/ionic-team/ionic-portals/pull/228))\r\n\r\n",
+    "mdBody": "### What's Changed\n* **fix**: Peer dep was not upgraded to cap 5 ([#228](https://github.com/ionic-team/ionic-portals/pull/228))\n",
     "body": "<h3>What's Changed</h3>\n<ul>\n<li><strong>fix</strong>: Peer dep was not upgraded to cap 5 (<a href=\"https://github.com/ionic-team/ionic-portals/pull/228\">#228</a>)</li>\n</ul>\n",
     "name": "0.8.1",
     "published_at": "May 19 2023",
@@ -12,10 +12,10 @@
     "version": "0.8.1"
   },
   {
-    "productTitle": "Portals Web",
+    "productTitle": "Portals Web Plugin",
     "pageUrl": "https://ionic.io/docs/portals/for-web/changelog",
     "repo": "ionic-team/ionic-portals",
-    "mdBody": "### What's Changed\r\n* **feat**: BREAKING! Overhauls to pub/sub ([#222](https://github.com/ionic-team/ionic-portals/pull/222))\r\n* **chore**: Update method for retrieving initial context on Android ([#226](https://github.com/ionic-team/ionic-portals/pull/226))\r\n\r\n",
+    "mdBody": "### What's Changed\n* **feat**: BREAKING! Overhauls to pub/sub ([#222](https://github.com/ionic-team/ionic-portals/pull/222))\n* **chore**: Update method for retrieving initial context on Android ([#226](https://github.com/ionic-team/ionic-portals/pull/226))\n",
     "body": "<h3>What's Changed</h3>\n<ul>\n<li><strong>feat</strong>: BREAKING! Overhauls to pub/sub (<a href=\"https://github.com/ionic-team/ionic-portals/pull/222\">#222</a>)</li>\n<li><strong>chore</strong>: Update method for retrieving initial context on Android (<a href=\"https://github.com/ionic-team/ionic-portals/pull/226\">#226</a>)</li>\n</ul>\n",
     "name": "0.8.0",
     "published_at": "May 17 2023",
@@ -24,10 +24,10 @@
     "version": "0.8.0"
   },
   {
-    "productTitle": "Portals Web",
+    "productTitle": "Portals Web Plugin",
     "pageUrl": "https://ionic.io/docs/portals/for-web/changelog",
     "repo": "ionic-team/ionic-portals",
-    "mdBody": "### What's Changed\r\n\r\n* **feat(plugin)**: Add `assetMaps` key to initialContext ([#209](https://github.com/ionic-team/ionic-portals/pull/209))\r\n\r\n\r\n\r\n",
+    "mdBody": "### What's Changed\n* **feat(plugin)**: Add `assetMaps` key to initialContext ([#209](https://github.com/ionic-team/ionic-portals/pull/209))\n",
     "body": "<h3>What's Changed</h3>\n<ul>\n<li><strong>feat(plugin)</strong>: Add <code>assetMaps</code> key to initialContext (<a href=\"https://github.com/ionic-team/ionic-portals/pull/209\">#209</a>)</li>\n</ul>\n",
     "name": "0.7.1",
     "published_at": "March 1 2023",
